refactor(api): move CW api to CW folder
Ref. metriport/metriport-internal#369
